Migrate website build config to Nextra 4 (Next 15, Tailwind v4)
Replace the Nextra 2 / Pages Router config with a Nextra 4 scaffold:
next.config.mjs (ESM, plausible proxy + redirects.js + .md rewrite preserved),
Tailwind v4 CSS-first globals.css (blue palette + primaryHue as CSS vars),
@tailwindcss/postcss, and pinned next@15 + zod@4.3.0 per the migration spike.
